feat(topology/algebra/module/character_space): Introduce the character space of an algebra (#12838)
The character space of a topological algebra is the subset of elements of the weak dual that are also algebra homomorphisms. This space is used in the Gelfand transform, which gives an isomorphism between a commutative C⋆-algebra and continuous functions on the character space of the algebra. This, in turn, is used to construct the continuous functional calculus on C⋆-algebras.



Co-authored-by: Frédéric Dupuis <31101893+dupuisf@users.noreply.github.com>