feat(algebra/group/to_additive): Let to_additive_doc linter ignore aliases
many of our `nolint.txt` entires are due to code of this shape:

    lemma foo := .. /- no docstring -/
    alias foo <- bar
    attribute [to_additive add_bar] bar

where now `bar` has a docstring (from `alias`), but `bar_add` does not.

This PR amkes `to_additive_doc_lint` recognize and ignore that case.

A later refinement could be that `alias` also copies the `to_additive`
attribute, but that’d be a bigger change (and I’m not sure what the
syntax should be).