feat(src/group_theory/free_group): is_free_group via `free_group X ≃* G`
the previous definition of the `is_free_group` class was defined via the universal
property of free groups, which is intelectually pleasing, but
technically annoying, due to the universe problems of quantifying over
“all other groups” in the definition. To work around them, many
definitions had to be duplicated.

This changes the definition o `is_free_group` to contian an isomorphism
between the `free_group` over the generator and `G`. It also moves this
class into `free_group.lean`, so that it can be found more easily.

The users of this API are updated.

Relevant Zulip thread:
<https://leanprover.zulipchat.com/#narrow/stream/113488-general/topic/universe.20levels.20for.20is_free_group>