docs(*): Wrap some links in < … >
I noticed that many docs say

    See https://stacks.math.columbia.edu/tag/001T.

and the our documentation will include the final `.` in the URL, causing
the URL to not work.

This tries to fix some of these instances. I intentionally applied this
to some URLs ending with a space, because it does not hurt to be
explicit, and the next contributor cargo-culting the URL is more likely
to get this right.

Obligatory xkcd reference: https://xkcd.com/208/