feat(topology/uniform_space/uniform_convergence): Uniform Cauchy sequences
A sequence of functions f_n is pointwise Cauchy if ∀ε ∀x, for sufficiently
large m and n we have |f_m x - f_n x| < ε. A sequence of functions is
_uniformly_ Cauchy if ∀x ∀ε, for sufficiently large m and n we have |f_m
x - f_n x| < ε.

As a sequence of functions is pointwise Cauchy if (and when the
underlying space is complete, only if) the sequence converges, a
sequence of functions is uniformly Cauchy if (and when the underlying
space is complete, only if) the sequence uniformly converges.

This notion is commonly used when to bootstrap convergence into uniform
convergence.