refactor(topology.metric_space.baire): add baire_space class and instances
Add a baire_space class containing the baire property (a countable intersection of open dense sets is dense).
The Baire category theorem for complete metric spaces becomes an instance of baire_space.
Previous consequences of the Baire property use baire_space as an hypothesis.
Add an additional instance of baire_space for locally compact spaces, in effect extending all the consequences of the Baire property to locally compact spaces.